test(club): merge ClubMapperUpdateMemberCountTest into ClubMapperTest
ClubMapperUpdateMemberCountTest's setUp() was a byte-for-byte duplicate of
ClubMapperTest's ~55-line DDL (oauth_users / clubs / club_social_medias /
club_member), and it stood up its own named in-memory database and its own
Spring context just to avoid interfering with the *shared* "hsclubs_test"
database used by other suites -- the concern 9b8368b actually addressed.
Both classes drop-and-recreate the exact same four tables for the exact same
reason (isolating club-mapper tests from OAuthUserMapperTest's differently
shaped oauth_users table), so they can safely share one database, one
fixture, and one Spring context instead of paying for a second of each.

Move updateIgnoresAClientSuppliedMemberCountInsteadOfOverwritingTheDerived-
Value() and insertIgnoresAClientSuppliedMemberCountAndLeavesTheColumnAt-
ItsDefault() into ClubMapperTest verbatim (assertions unchanged) and delete
ClubMapperUpdateMemberCountTest. ClubMapperTest now runs 4 tests against a
single "club_mapper_test" H2 instance instead of splitting 2+2 across two
databases and two contexts.
